requirements.txt
Flask==2.0.2
google-cloud-speech==2.7.0
google-cloud-texttospeech==2.5.1
openai==0.27.0
gunicorn==20.1.0
